Update Start up text.py
I added a confirmation screen when selecting the password generator option in the main menu. This means that the user won't accidentally start the generation process.
To make everything look more streamlined and nice to look at. I added some print("\n") statements through the code so that when running the text is spread out and easy to read